Move tests to have same shape as webapp
