+++ tests/publisher/snaps/tests_post_binary_metadata.py
@@ -2,7 +2,7 @@ import unittest
 import responses
 import json
 import io
-from tests.endpoint_testing import BaseTestCases
+from tests.publisher.endpoint_testing import BaseTestCases
 
 
 class PostBinaryMetadataListingPageLoggedOut(
