+++ tests/publisher/snaps/tests_post_listing.py
@@ -2,7 +2,7 @@ import unittest
 import responses
 import json
 
-from tests.endpoint_testing import BaseTestCases
+from tests.publisher.endpoint_testing import BaseTestCases
 
 
 class PostListingPageNotAuth(BaseTestCases.EndpointLoggedOut):
