+++ tests/publisher/snaps/tests_revision.py
@@ -1,6 +1,6 @@
 import responses
 
-from tests.endpoint_testing import BaseTestCases
+from tests.publisher.endpoint_testing import BaseTestCases
 
 
 class RevisionHistoryPageNotAuth(BaseTestCases.EndpointLoggedOut):
