+++ tests/publisher/tests_account_details.py
@@ -1,6 +1,6 @@
 import unittest
 import responses
-from tests.endpoint_testing import BaseTestCases
+from tests.publisher.endpoint_testing import BaseTestCases
 
 
 # Make sure tests fail on stray responses.
