+++ tests/publisher/tests_publisher_logic.py
