+++ tests/snapcraft/tests_public.py
