+++ tests/store/tests_public_logic.py
